Create Connect4 game in java
For this Java coding to create this Connect 4 game, mainly two classes are created, one is Main and another is TestDrawTheGame. The Main class is used for making the game and TestDrawTheGame class used for the unit test of five methods present in the Main class.

In the Main class, a 2D array is created where the row number is 6 and the column number is 7. These two are set using the final static variable named ROW and COLUMN.